test: fix tests.
